AudioDataset: modify its constructor to accepts example_dict iterator
..., instead of accepting the corpus path. This facilitates later preprocess
sharding work.
